requirements.txt
numpy
scipy
PyYAML
mujoco>=3.3.0
pytest
